feat: implementação completa da API REST Back End Java Sênior

+++ src/main/java/com/example/musicapi/repository/RegionalRepository.java
@@ -1,10 +1,13 @@
 package com.example.musicapi.repository;
+import java.util.stream.Collectors;
 
 import com.example.musicapi.model.RegionalEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface RegionalRepository extends JpaRepository<RegionalEntity, Integer> {
+public interface RegionalRepository extends JpaRepository<RegionalEntity, Long> {
+
     List<RegionalEntity> findByAtivoTrue();
+
 }
